requirements.txt
--prefer-binary
fastapi==0.115.0
uvicorn==0.31.0
python-multipart==0.0.9
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=2.0.0
Pillow>=10.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Additional utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0

